refactor: home action buttons
