+++ src/app/pages/home/components/home-action-button.tsx
@@ -3,7 +3,7 @@ import { memo } from 'react';
 import { Box, ButtonProps, Text } from '@stacks/ui';
 
 interface HomeActionButtonProps extends ButtonProps {
-  icon: any;
+  icon?: any;
   label: string;
   buttonComponent(props: ButtonProps): React.JSX.Element;
 }
@@ -11,8 +11,8 @@ export const HomeActionButton = memo((props: HomeActionButtonProps) => {
   const { icon, label, buttonComponent: Button, ...rest } = props;
 
   return (
-    <Button height="36px" px="base-tight" py="tight" position="relative" {...rest}>
-      <Box as={icon} mr="tight" size="14px" />
+    <Button height="36px" minWidth="80px" px="base-tight" py="tight" position="relative" {...rest}>
+      {icon ? <Box as={icon} mr="tight" size="14px" /> : null}
       <Text fontSize="14px">{label}</Text>
     </Button>
   );
