--- src/app/pages/swap/index.ts
@@ -1,2 +1,0 @@
-export * from './swap';
-export * from './swap-routes';
