+++ src/app/pages/swap/swap-container.tsx
@@ -2,14 +2,13 @@ import { Outlet } from 'react-router-dom';
 
 import { Flex } from '@stacks/ui';
 
-import { whenPageMode } from '@app/common/utils';
-import { ModalHeader } from '@app/components/modal-header';
 import { useRouteHeader } from '@app/common/hooks/use-route-header';
-
+import { whenPageMode } from '@app/common/utils';
 import { CENTERED_FULL_PAGE_MAX_WIDTH } from '@app/components/global-styles/full-page-styles';
+import { ModalHeader } from '@app/components/modal-header';
 
 export function SwapContainer() {
-  useRouteHeader(<ModalHeader hideActions defaultGoBack title="Swap" />, true);
+  useRouteHeader(<ModalHeader defaultGoBack hideActions title="Swap" />, true);
 
   return whenPageMode({
     full: (
