+++ src/app/pages/swap/swap.routes.tsx
